roboflow-function/requirements.txt
google-cloud-storage
inference-sdk